First lighting commit(broken)

+++ core/src/com/upa/gun/SpawnIndicator.java
@@ -3,13 +3,13 @@ package com.upa.gun;
 import com.badlogic.gdx.physics.box2d.World;
 
 public class SpawnIndicator {
-    int x, y;
+    float x, y;
     float timeElapsed;
     float timeUntilSpawn;
     Class spawnType;
     boolean markedForDeletion;
 
-    SpawnIndicator(int x, int y, float timeElapsed, float timeUntilSpawn, Class spawnType) {
+    SpawnIndicator(float x, float y, float timeElapsed, float timeUntilSpawn, Class spawnType) {
         this.x = x;
         this.y = y;
         this.timeElapsed = timeElapsed;
